% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Read in files and create visual of all accidents in state for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer - the FIPS code for a US state}

\item{year}{a four digit integer - the year of accidents you are interested in}
}
\value{
a visual of all the accidents in state for a given year by geographic location
}
\description{
Read in files and create visual of all accidents in state for a given year
}
\note{
This function will not work as intended if an \code{year} is not an integer or \code{state.num} is not an integer from 1-56
}
