% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a CSV file and convert it into a data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string - The name of a csv file to import}
}
\value{
A data frame containing the information from the imported csv
}
\description{
Read in a CSV file and convert it into a data frame
}
\note{
This function will error out if the \code{filename} does not exist
}
\examples{
fars_read("accident_2013.csv.bz2")
}
