% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in files and create tibble of month and year combinations in each file}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of four digit integers - all the years of files you are interested in}
}
\value{
a list of tibbles containing month/year combinations in the files specified by the \code{years} argument
}
\description{
Read in files and create tibble of month and year combinations in each file
}
\note{
This function will not work as intended if an \code{years} is not an integer or a vector of integers
}
\examples{
fars_read_years(c(2013,2014))
}
