% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read in files and create tibble of counts of year/month combinations in all imported files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of four digit integers - all the years of files you are interested in}
}
\value{
a tibble containing counts of month/year combinations in the files specified by the \code{years} argument
}
\description{
Read in files and create tibble of counts of year/month combinations in all imported files
}
\note{
This function will not work as intended if an \code{years} is not an integer or a vector of integers
}
\examples{
fars_summarize_years(c(2013,2014))
}
