% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name for specific accident year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer - the year of accidents you are interested in}
}
\value{
the name of a specific file to import
}
\description{
Create file name for specific accident year
}
\note{
This function will not work as intended if an \code{year} is not an integer
}
\examples{
make_filename(2013)
}
